migrating misskey feature.

+++ core/src/commonMain/kotlin/work/socialhub/planetlink/micro/MicroBlogComment.kt
@@ -11,8 +11,8 @@ open class MicroBlogComment(
     var liked: Boolean = false
     var shared: Boolean = false
 
-    var likeCount: Long? = null
-    var shareCount: Long? = null
+    var likeCount: Int? = null
+    var shareCount: Int? = null
 
     var replyTo: Identify? = null
 }
