+++ core/src/commonMain/kotlin/work/socialhub/planetlink/model/Reaction.kt
@@ -9,7 +9,7 @@ class Reaction {
     var name: String? = null
     var emoji: String? = null
     var iconUrl: String? = null
-    var count: Long? = null
+    var count: Int? = null
 
     /** 認証ユーザーがリアクションしたか？  */
     var reacting: Boolean = false
